Some changes happened

+++ customer/customer/src/main/java/com/customer/customer/entity/Customer.java
@@ -1,12 +1,9 @@
 package com.customer.customer.entity;
 
 
+
 import jakarta.persistence.*;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.Max;
-import jakarta.validation.constraints.Min;
 import lombok.*;
-import org.hibernate.annotations.GenericGenerator;
 
 import java.util.List;
 
@@ -18,13 +15,13 @@ import java.util.List;
 public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    public Long customerId;
+    public int customerId;
 
     public String fullName;
     public Integer age;
     public String gender;
     public String mobileNumber;
     public String email;
+//    transient public List<AddresssClient> addresssClients;
 
-//    transient private List<CustomerWithAddress> customerWithAddresses;
 }
